Move FormRepositoryTest into forms
