+++ collect_app/src/test/java/org/odk/collect/android/database/DatabaseFormsRepositoryTest.java
@@ -6,7 +6,7 @@ import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.database.forms.DatabaseFormsRepository;
-import org.odk.collect.android.forms.FormsRepositoryTest;
+import org.odk.collect.forms.testsupport.FormsRepositoryTest;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.storage.StorageInitializer;
 import org.odk.collect.android.storage.StoragePathProvider;
