+++ collect_app/src/test/java/org/odk/collect/android/forms/InMemFormsRepositoryTest.java
@@ -3,6 +3,7 @@ package org.odk.collect.android.forms;
 import org.junit.Before;
 import org.odk.collect.android.support.InMemFormsRepository;
 import org.odk.collect.forms.FormsRepository;
+import org.odk.collect.forms.testsupport.FormsRepositoryTest;
 import org.odk.collect.testshared.TempFiles;
 
 import java.util.function.Supplier;
